added tests

+++ owlbear/p_items/BasicItem.cs
@@ -10,10 +10,10 @@ namespace Items
     {
         private string  m_name;
         private string  m_desc;
-        private int     m_weight;
-        private int     m_value;
+        private double  m_weight;
+        private double  m_value;
 
-        public BasicItem(string name, string desc, int weight, int value)
+        public BasicItem(string name, string desc, double weight, double value)
         {
             m_name   = name;
             m_desc   = desc;
@@ -23,8 +23,8 @@ namespace Items
 
         public string Name  { get => m_name;   set => m_name   = value; }
         public string Desc  { get => m_desc;   set => m_desc   = value; }
-        public int Weight   { get => m_weight; set => m_weight = value; }
-        public int Value    { get => m_value;  set => m_value  = value; }
+        public double Weight   { get => m_weight; set => m_weight = value; }
+        public double Value    { get => m_value;  set => m_value  = value; }
 
     }
 }
